pyproject.toml
[tool.poetry]
name = "oz-collabo-project"
version = "0.1.0"
description = ""
authors = ["Donggurii._.dd <ehdgur741@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
psycopg = "^3.2.3"
black = "^24.10.0"
isort = "^5.13.2"
coverage = "^7.6.4"
gunicorn = "^23.0.0"
django-redis = "^5.4.0"
channels-redis = "^4.2.0"
django-storages = "^1.14.4"
pillow = "^11.0.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
requests = "^2.32.3"
django-extensions = "^3.2.3"
django-multiselectfield = "^0.1.13"
faker = "^33.0.0"
daphne = "^4.1.2"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
boto3 = "^1.35.76"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120